Refactor the modules and folders (sep of concern)
