+++ app/src/main/java/com/zip_feast/presentation/auth/authnavigation/Screen.kt
@@ -1,4 +1,4 @@
-package com.zip_feast.utils.authnavigation
+package com.zip_feast.presentation.auth.authnavigation
 
 sealed class Screen(val route: String) {
     object Login : Screen("login")
