+++ app/src/main/java/com/zip_feast/presentation/auth/screens/UserRegistrationScreeen.kt
@@ -54,8 +54,8 @@ import com.zip_feast.data.remote.models.UserRequest
 import com.zip_feast.presentation.theme.Black
 import com.zip_feast.presentation.theme.BlueGray
 import com.zip_feast.presentation.theme.Roboto
-import com.zip_feast.utils.authnavigation.Screen
-import com.zip_feast.viewmodels.auth.AuthViewModel
+import com.zip_feast.presentation.auth.authnavigation.Screen
+import com.zip_feast.presentation.auth.authviewmodels.AuthViewModel
 
 
 @Composable
