+++ app/src/main/java/com/zip_feast/presentation/dashboard/navigations/NavGraph.kt
@@ -12,8 +12,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import com.zip_feast.presentation.dashboard.navigations.navmodel.ProductDetail
-import com.zip_feast.presentation.dashboard.screens.AccountScreen
-import com.zip_feast.presentation.dashboard.screens.CartScreen
+import com.zip_feast.presentation.profile.Screens.AccountScreen
+import com.zip_feast.presentation.cart.Screens.CartScreen
 import com.zip_feast.presentation.dashboard.screens.ExploreScreen
 import com.zip_feast.presentation.dashboard.screens.HomeScreen
 import com.zip_feast.presentation.dashboard.screens.ServicesScreen
