+++ app/src/main/java/com/zip_feast/presentation/profile/Screens/AccountScreen.kt
@@ -1,15 +1,12 @@
-package com.zip_feast.presentation.dashboard.screens
+package com.zip_feast.presentation.profile.Screens
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.navigation.NavHostController
 
 @Composable
 fun AccountScreen() {
